dev-requirements.txt
pytest==7.1.3
pytest-cov==4.0.0
requests==2.32.4
aiohttp>=3.0.0
asyncio
websocket-client